Move set::tests to its own file
